requirements.txt
python-telegram-bot==20.7
Pillow>=9.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
SQLAlchemy>=1.4.0
pytz>=2022.0
